Optionally merge prefix with host part when creating IPv6 address
Previous behavior allowed for either randomizing the host part or
filling it with zeros. An IPv6 address with the host part set to all
zeros is reserved for router anycasts by RFC4291. Instead, we're
allowing the caller to either specify the host part of the address
or we randomize it if NULL.